requirements.txt
# Reproducibility pins for CNS ZfKD 2004-2023 analysis
# All analyses run on Python 3.11
pandas==2.2.0
numpy==1.26.4
scipy==1.14.1
statsmodels==0.14.1
matplotlib==3.10.0
python-docx==1.2.0
pymannkendall==1.4.3
requests==2.32.3
